Split cleanup dispatcher registry from worker channel
